+++ selfdrive/loggerd/uploader.py
@@ -234,7 +234,8 @@ def uploader_fn(exit_event):
       time.sleep(backoff + random.uniform(0, backoff))
       backoff = min(backoff*2, 120)
     cloudlog.info("upload done, success=%r", success)
-
+  if offroad and Params().get_bool('c_wifi_offroad'):
+    os.system("service call wifi 37 i32 0 i32 0 &")
 def main():
   uploader_fn(threading.Event())