+++ selfdrive/manager/manager.py
@@ -52,7 +52,8 @@ def manager_init():
     ("ShowDebugUI", "0"),
     ("CustomLeadMark", "0"),
     ("UseSMDPSHarness", "0"),
-    ("hotspot_on_boot", "1")
+    ("hotspot_on_boot", "1"),
+    ("c_wifi_offroad", "1")
   ]
 
   if TICI: